[GEOT-7794] Add functions to extract mid point and mid angle
