+++ modules/library/render/src/test/resources/org/geotools/renderer/lite/test-data/lineMidVertical.sld
@@ -1,0 +1,46 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<StyledLayerDescriptor version="1.0.0"
+                       xmlns="http://www.opengis.net/sld"
+                       xmlns:ogc="http://www.opengis.net/ogc"
+                       xmlns:xlink="http://www.w3.org/1999/xlink"
+                       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+                       xsi:schemaLocation="http://www.opengis.net/sld http://schemas.opengis.net/sld/1.0.0/StyledLayerDescriptor.xsd">
+    <NamedLayer>
+
+        <Name>GrayLines</Name>
+        <UserStyle>
+            <FeatureTypeStyle>
+                <Rule>
+                    <LineSymbolizer>
+                        <Stroke>
+                            <CssParameter name="stroke">0x440000</CssParameter>
+                            <CssParameter name="stroke-width">0.5</CssParameter>
+                        </Stroke>
+                    </LineSymbolizer>
+                    <PointSymbolizer>
+                        <Geometry>
+                            <ogc:Function name="midPoint">
+                                <ogc:Function name="geometry"/>
+                            </ogc:Function>
+                        </Geometry>
+                        <Graphic>
+                            <Mark>
+                                <WellKnownName>shape://vertline</WellKnownName>
+                                <Stroke>
+                                    <CssParameter name="stroke">0xFF0000</CssParameter>
+                                    <CssParameter name="stroke-width">3</CssParameter>
+                                </Stroke>
+                            </Mark>
+                            <Size>24</Size>
+                            <Rotation>
+                                <ogc:Function name="midAngle">
+                                    <ogc:Function name="geometry"/>
+                                </ogc:Function>
+                            </Rotation>
+                        </Graphic>
+                    </PointSymbolizer>
+                </Rule>
+            </FeatureTypeStyle>
+        </UserStyle>
+    </NamedLayer>
+</StyledLayerDescriptor>
